Reorganize DB code
