+++ linera-indexer/lib/src/db/sqlite/consts.rs
