+++ linera-indexer/lib/src/lib.rs
@@ -19,13 +19,7 @@ pub mod rocks_db;
 #[cfg(feature = "scylladb")]
 pub mod scylla_db;
 
-pub mod database_trait;
-#[cfg(test)]
-pub mod mock_database;
-pub mod sqlite;
-pub mod sqlite_db;
-pub mod sqlite_runner;
-
+pub mod db;
 pub mod grpc;
 
 // Generated protobuf types
